Group `font-weight` examples
